feat: Add license file integration, refine CASC bundle loading with local storage, introduce a webtop browser service, and parameterize controller names.

+++ up.sh
@@ -10,17 +10,26 @@ else
   echo "JAVA_HOME is set to: $JAVA_HOME"
 fi
 
+if [[ ! -f "./license.crt" || ! -f "./license.key" ]]; then
+  echo "CloudBees CI license files not found. Create them or copy them to the current directory"
+  exit 1
+fi
+
+
+
 if [[ ! -f "./ssl/server.crt" || ! -f "./ssl/server.key" ]]; then
   ./generate-ssl-cert.sh
 else
   echo "SSL certificates already exist. Run 'rm -Rf ssl/*' to recreate them"
 fi
 ./down.sh
-envsubst < haproxy-config/haproxy-ssl.cfg > "${HA_PROXY_CONFIG}"
 
+# Create HAProxy config from template
+envsubst < haproxy-config/haproxy-ssl.cfg > "${HA_PROXY_CONFIG}"
 
 docker-compose  up -d --build
 
-#open http://127.0.0.1:6080
+#open https://cjoc.local
+#open https://controller.local
 
 #docker-compose logs -f
